refactor(metrics): extract FullSyncResult and LiveCatchupResult to .types files
