+++ services/outlook-semantic-mcp/src/features/sync/full-sync/full-sync.types.ts
@@ -1,0 +1,5 @@
+export type FullSyncResult =
+  | { status: 'skipped'; reason: string }
+  | { status: 'waiting-for-ingestion' }
+  | { status: 'completed' }
+  | { status: 'failed'; error: unknown };
