+++ services/outlook-semantic-mcp/src/features/sync/live-catch-up/live-catch-up.command.ts
@@ -30,12 +30,11 @@ import {
   graphMessagesResponseSchema,
 } from '../../process-email/dtos/microsoft-graph.dtos';
 import { ProcessEmailCommand } from '../../process-email/process-email.command';
+import type { LiveCatchupResult } from './live-catch-up.types';
 
 export const RUNNING_LIVE_CATCHUP_THRESHOLD_MINUTES = 20;
 export const FAILED_LIVE_CATCHUP_THRESHOLD_MINUTES = 5;
 export const READY_LIVE_CATCHUP_THRESHOLD_MINUTES = 30;
-
-type LiveCatchupResult = { status: 'completed' | 'skipped' } | { status: 'failed'; err: unknown };
 
 @Injectable()
 export class LiveCatchUpCommand {
