+++ services/outlook-semantic-mcp/src/features/sync/live-catch-up/live-catch-up.types.ts
@@ -1,0 +1,3 @@
+export type LiveCatchupResult =
+  | { status: 'completed' | 'skipped' }
+  | { status: 'failed'; err: unknown };
